Projectiles Update
I have changed the projectiles so that they are all spawned through the character controller class via a blueprint and a Projectile Movement component added there. The projectile movement component sets the projectiles velocity to a negative value and multiplies it by the moveright axis. They are being shot in the correct direction.

I still need to add in the deleting of the projectiles and set it up so that the player can not shoot projectiles if they are facing the x axis. I also need to add the players velocity to the projectiles so they move faster than the player.